Remove hello world, .gitignore, and binary file.
Trying to ignore bazel files is kind of silly because ideally I should
be able to clone this repo on any computer and resume work on it,
without needing to set up bazel again.

Uploading the binary file also seems silly because I would never run
the binary without building it first.